Switch from double to typedef Real.
This will make it easy to change precision to float,
if this is deemed appropriate. Per now, changing to float
gives no benefit in terms of speed. It is actually slightly worse,
indicating that maybe some armadillo things converts back and forth.

When vectorisation optimisation starts, it might be beneficial to use
a float, as more floats can fit into the wide registers.