Add factor 2 that was missing.
With this change we seem to always get an acceptance prob
for new moves of exaclty 1. This is only without interaction though,
and seems to exact to be by coincidence or bug. I believe this to be due
to some properties of the simplicity of the non-interacting
wavefunction. The green function part seems to always be the inverse of
|Psi_new|^2/|Pse_old|^2.

Might be an error, but might also just be optimal sampling with 100%
acceptance of new moves.