Implement memorized distance lookup.
Depends on detecting when bosons are changed. This required
the ability to distinguish between modifier access and const access,
beyond simply relying on const variables. This is because some routines
will both modify the system, and query for the distance.

Solution:

Use `operator()` when _modifying_, and `operator[]` for const access.

Then the distance lookup is able to mark modified elements, and
recalculate the distances as needed.