Bugfix: Avoid NaN when initializing samplers.
The previous state wavefunction evaluation should not be based on the
system that is used to initialize the sampler classes, but rather the
system after `initialize_system` has set up the internal system config.
The reason is that the typical usage pattern is to send a uninitialized
system in to the constructor, only as away to define the system
dimensions. That system might be zero initialized (which can be bad in
some cases), or it might be uninitialized with all the random trouble
that comes along with that.

Just as an extra check, NaNs are explicitly detected (nice debug hook).