Separate backend and Python package.
This allows to make a package in which there can be both C++ bindings
_and_ regular Python code within the same modules.

Say I want to add a function to the `qflow.wavefunction` submodule. Then
I add a can add a file `qflow/wavefunction.py` and write `from
_qflow_backend.wavefunction import *` at the top and then add everything
else. Or I can make a folder `qflow/wavefunction/` and write this in the
`qflow/wavefunction/__init__.py` file.