Revert "omp parallelize sum pooling"
This reverts commit e0e1c3dbd204853e9474daa8f4bb2b44f08f9cfd.

The parallelization of SumPooling cannot be implemented at present,
because some wavefunctions are not thread safe. This refers as of now
only to Dnn, and is because each layer has a stored state that is
changed on calls. Having multiple threads calling the evaluation leads
to bad things (malloc errors at runtime).

If this is to be reintroduced, it's likely that each thread would need a
copy of the wavefunction. This may, however, mean significant copying
work that might only be worth it for very large N. Not to mention the
issues of making a copy of an abstract base class pointer...