Fix distance caching API + mini benchmark
Benchmark indicates slight speed decrease for the benchmark case. This
indicates that caching should only be done for
wavefunctions/hamiltonians that makes heavy use of the distances. The
bench case is the simple gaussian, which uses none of this, and only the
mean distance computation is used (which is one distance eval per
iteration, so a really bad benchmark for this ).