BREAKING CHANGE: fix error in blocking
An off-by-1 error in the adapted blocking code lead
to qflow calculating statistics on data that underwent
one too many blocking iterations.

The result of this error was that standard errors came out slightly
larger than they should have been. After this fix the SEM will be
somewhat smaller in general. The reason is that one more blocking
iteration equates to measurements that have more auto-correlation.
The old blocking code effectively over-corrected.

This should therefore not put any findings in the thesis into question,
as results that where significant before should only be more significant
now.

BREAKING CHANGE: Using this new version of the code will not reproduce
all results exactly. If exactly reproducing the results is the goal, the
git tag from when the thesis was submitted should be used.